requirements.txt
Django==3.1.3
djangorestframework==3.12.2
psycopg2
gunicorn==19.9.0